+++ clients/derivatives-trading-usds-futures/src/rest-api/types/user-commission-rate-response.ts
@@ -37,4 +37,10 @@ export interface UserCommissionRateResponse {
      * @memberof UserCommissionRateResponse
      */
     takerCommissionRate?: string;
+    /**
+     *
+     * @type {string}
+     * @memberof UserCommissionRateResponse
+     */
+    rpiCommissionRate?: string;
 }