+++ clients/derivatives-trading-usds-futures/src/websocket-streams/types/algo-update-o.ts
@@ -163,4 +163,10 @@ export interface AlgoUpdateO {
      * @memberof AlgoUpdateO
      */
     gtd?: number | bigint;
+    /**
+     *
+     * @type {string}
+     * @memberof AlgoUpdateO
+     */
+    rm?: string;
 }