+++ clients/derivatives-trading-usds-futures/tests/unit/rest-api/AccountApiTest.test.ts
@@ -2331,6 +2331,7 @@ describe('AccountApi', () => {
                     symbol: 'BTCUSDT',
                     makerCommissionRate: '0.0002',
                     takerCommissionRate: '0.0004',
+                    rpiCommissionRate: '0.00005',
                 })
             );
 
@@ -2359,6 +2360,7 @@ describe('AccountApi', () => {
                     symbol: 'BTCUSDT',
                     makerCommissionRate: '0.0002',
                     takerCommissionRate: '0.0004',
+                    rpiCommissionRate: '0.00005',
                 })
             );